Move create_ctx_with{,out}_session from tests to common module
This ensures that all the tests initialize the context and TCTI in a
consistent way, and TEST_TCTI is useful throughout the tests.
