+++ tests/abstraction_nv_tests.rs
@@ -1,51 +1,17 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{convert::TryFrom, env, str::FromStr};
+use std::convert::TryFrom;
 use tss_esapi::abstraction::nv;
 use tss_esapi::{
-    constants::{
-        algorithm::{Cipher, HashingAlgorithm},
-        tss::*,
-        types::session::SessionType,
-    },
+    constants::algorithm::HashingAlgorithm,
     handles::NvIndexTpmHandle,
     nv::storage::{NvAuthorization, NvIndexAttributes, NvPublicBuilder},
     structures::MaxNvBuffer,
-    utils::TpmaSessionBuilder,
-    Context, Tcti,
 };
 
-pub fn create_ctx_without_session() -> Context {
-    let tcti = match env::var("TEST_TCTI") {
-        Err(_) => Tcti::Mssim(Default::default()),
-        Ok(tctistr) => Tcti::from_str(&tctistr).expect("Error parsing TEST_TCTI"),
-    };
-    unsafe { Context::new(tcti).unwrap() }
-}
-
-fn create_ctx_with_session() -> Context {
-    let mut ctx = create_ctx_without_session();
-    let session = ctx
-        .start_auth_session(
-            None,
-            None,
-            None,
-            SessionType::Hmac,
-            Cipher::aes_256_cfb(),
-            HashingAlgorithm::Sha256,
-        )
-        .unwrap();
-    let session_attr = TpmaSessionBuilder::new()
-        .with_flag(TPMA_SESSION_DECRYPT)
-        .with_flag(TPMA_SESSION_ENCRYPT)
-        .build();
-    ctx.tr_sess_set_attributes(session.unwrap(), session_attr)
-        .unwrap();
-    ctx.set_sessions((session, None, None));
-
-    ctx
-}
+mod common;
+use common::create_ctx_with_session;
 
 #[test]
 fn read_full() {
