+++ tests/structures_capabilitydata.rs
@@ -1,19 +1,13 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{env, str::FromStr};
-use tss_esapi::{constants::types::capability::CapabilityType, Context, Tcti};
+use tss_esapi::constants::types::capability::CapabilityType;
+
+mod common;
+use common::create_ctx_without_session;
 
 mod test_capabs {
     use super::*;
-
-    pub fn create_ctx_without_session() -> Context {
-        let tcti = match env::var("TEST_TCTI") {
-            Err(_) => Tcti::Mssim(Default::default()),
-            Ok(tctistr) => Tcti::from_str(&tctistr).expect("Error parsing TEST_TCTI"),
-        };
-        unsafe { Context::new(tcti).unwrap() }
-    }
 
     #[test]
     fn test_algorithms() {
