+++ tests/utils_tests.rs
@@ -1,16 +1,10 @@
 // Copyright 2020 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{env, str::FromStr};
-use tss_esapi::{utils, Context, Tcti};
+use tss_esapi::utils;
 
-pub fn create_ctx_without_session() -> Context {
-    let tcti = match env::var("TEST_TCTI") {
-        Err(_) => Tcti::Mssim(Default::default()),
-        Ok(tctistr) => Tcti::from_str(&tctistr).expect("Error parsing TEST_TCTI"),
-    };
-    unsafe { Context::new(tcti).unwrap() }
-}
+mod common;
+use common::create_ctx_without_session;
 
 #[test]
 fn get_tpm_vendor() {
